Grill timing mechanism now working

+++ JavaDev/core/src/com/mygdx/game/Ingredient.java
@@ -42,7 +42,7 @@ public class Ingredient {
         return false;
     }
 
-    public float getCookingStartTime() {
+    public long getCookingStartTime() {
         return 0;
     }
 }
